update text sizes and spacing
